Don't autofocus the editable constraint field.
There can be any number of these on a page, so setting autofocus here is
a Bad Idea (TM). It's probably a holdover from when the input was an
accordion and we wanted to give you focus when you opened it (we do a
similar thing for the popover, for instance).

This property will cause you to focus (and potentially scroll) to the
last constraint on the page.