fix: hide project archive in OSS (#10004)
https://linear.app/unleash/issue/2-3569/fix-hide-project-archive-in-oss

Hides "project archive" in OSS.

I believe this is a bug. OSS only has one project and the project
archive was acting unexpectedly anyways since it was showing the same
default project as being archived. This is because in OSS we use the OSS
project-controller, not the Enterprise version (override) of it.