Feat(1-3801)/add more data to lifecycle tiles (#10092)
This is the first pass at the full lifecycle tiles. It adds the tile
header and current and historical median data.

I have also added large number handling to all the number instances in
the tile: in the header, the graph, and the median data. In doing so, I
exposed the algorithm we use in the PrettifyLargeNumber component.
Returning a react component isn't always a valid option (such as in the
chart). This does mean that you don't get a tooltip when you use the
function directly, but in things like the chart and the median
measurement that makes sense to me.

I've decided to return "No data" if the median days value is 0 or lower.

There's no data for historical medians yet, so I'm using the same number
for now.

<img width="1538" alt="image"
src="https://github.com/user-attachments/assets/72e6a90a-6b84-47ce-af02-59596a7ff91f"
/>