chore: fix skeleton loading indicator colors for lifecycle trend tiles (#10109)
Updates the skeleton loading colors for the lifecycle trend tiles
because the previous color was invisible on the median stats section in
light mode. You also wouldn't see any movement in the total number of
flags.

I have not touched the original skeleton color assignment because that
might have unintended consequences throughout the app. Also, I'm not
aware of a way (at the moment) that would do this automatically when
necessary. As such, I have only touched the loading indicators in the
lifecycle tile.

## Before:

Light mode:

![image](https://github.com/user-attachments/assets/3731eb0b-4d63-4622-8b20-0518e6266611)

Dark mode:

![image](https://github.com/user-attachments/assets/6df0197f-7283-4709-b369-38025011c879)


## After:

Light mode:

![image](https://github.com/user-attachments/assets/09f33fa8-322c-4a55-a6e8-4ef384104d74)

Dark mode:

![image](https://github.com/user-attachments/assets/60499b46-c9c7-49cc-a210-50d330682075)