Chore(1 3829)/cleanup pt 3 (#10133)
This PR removes more constraint inputs and validators that are not in
use anymore. Additionally, the old constraint components that are still
being used by the project action filter item, have been moved to that
directory. This also goes for ResolveInput which has been simplified to
only the inputs and operators used by actions filter item.

I've done a manual side-by-side comparison of the old and newly
refactored filter item, and it appears to be working the exact same.