chore: Use fixed-width sidebar instead of dynamic modal.
Uses a fixed-width sidebar component instead of the dynamic sidebar
component for the change request sidebar. This fixes a case where the
modal would suddenly grow narrower when a change was sent to review:

Before submitting (in main)

After submission (in main)

Before submitting (on this branch)

After submisson (on this branch)

I don't see any reason why the CR sidebar should be dynamic, so making
it fixed width with the solution we already have seems pretty sensible
to me. Keeps things consistent and prevents us from solving the same
problem multiple times in multiple ways.