feat: Timeline for cancelled CRs
Shows the cancelled state in the timeline for cancelled CRs.

Also, if timestamps are present, it will dynamically show the steps we have times for for rejected and cancelled CRs.

CRs can be cancelled after:
- In review
- Approved

CRs can be rejected after:
- In review
- Approved
- Schedule

For other states, use the existing steps. Cancelled and Rejected are both terminal states, so there's no future steps to show.

Regardless of what we have timestamps for, always show 'Draft', 'In review' and the final state (Rejected | Cancelled).